add hosts methods

+++ app/Http/Controllers/Api/HostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Models\HostsModel;
+use Illuminate\Support\Facades\Validator;
 
 class HostsController extends BaseController
 {
@@ -18,27 +19,42 @@ class HostsController extends BaseController
 
     public function store()
     {
-
+        $posts    = collect(request()->json())->toArray();
+        $validate = Validator::make($posts, [
+            'name' => "required",
+            'host' => "required"
+        ]);
+        if ($validate->fails()) {
+            return $this->error($validate->errors()->first(), 30002);
+        }
+        $insert = $this->model->create($posts)->save();
+        if (false === $insert) {
+            return $this->error('add error', 30004);
+        }
+        return $this->success();
     }
 
 
     public function index()
     {
-        return $this->succes($this->model->all());
+        return $this->success($this->model->all());
     }
 
     public function show(int $id)
     {
         $data = $this->model->find($id) ?: [];
-        return $this->succes($data);
+        return $this->success($data);
     }
 
     /**
      * @api DELETE /hosts/:id
      */
-    public function delete(int $id)
+    public function destroy(int $id)
     {
+        if (!$this->model->find($id)) {
+            return $this->error('host is not found', 30001);
+        }
         $res = $this->model->where('id', $id)->delete();
-        return response(['data' => [], 'code' => 20000]);
+        return $this->success([], 'ok');
     }
 }
